Applying suggestions and use ClassValue
