+++ dd-java-agent/instrumentation/mule-4/src/main/java/datadog/trace/instrumentation/mule4/JpmsMuleInstrumentation.java
@@ -5,10 +5,8 @@ import static net.bytebuddy.matcher.ElementMatchers.isConstructor;
 import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.agent.tooling.InstrumenterModule;
+import datadog.trace.agent.tooling.muzzle.Reference;
 import datadog.trace.api.Platform;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.implementation.bytecode.assign.Assigner;
-import org.mule.runtime.tracer.api.EventTracer;
 
 @AutoService(InstrumenterModule.class)
 public class JpmsMuleInstrumentation extends InstrumenterModule.Tracing
@@ -39,20 +37,23 @@ public class JpmsMuleInstrumentation extends InstrumenterModule.Tracing
   }
 
   @Override
-  public void methodAdvice(MethodTransformer transformer) {
-    // it does not work with typeInitializer()
-    transformer.applyAdvice(isConstructor(), getClass().getName() + "$JpmsClearanceAdvice");
+  public Reference[] additionalMuzzleReferences() {
+    return new Reference[] {
+      // added in 4.5.0
+      new Reference.Builder("org.mule.runtime.tracer.api.EventTracer")
+          .withMethod(
+              new String[0],
+              Reference.EXPECTS_NON_STATIC | Reference.EXPECTS_PUBLIC,
+              "endCurrentSpan",
+              "V",
+              "Lorg/mule/runtime/api/event/Event;")
+          .build(),
+    };
   }
 
-  public static class JpmsClearanceAdvice {
-    @Advice.OnMethodExit(suppress = Throwable.class)
-    public static void openOnReturn(@Advice.This(typing = Assigner.Typing.DYNAMIC) Object self) {
-      JpmsAdvisingHelper.allowAccessOnModuleClass(self.getClass());
-    }
-
-    private static void muzzleCheck(final EventTracer<?> tracer) {
-      // introduced in 4.5.0
-      tracer.endCurrentSpan(null);
-    }
+  @Override
+  public void methodAdvice(MethodTransformer transformer) {
+    // it does not work with typeInitializer()
+    transformer.applyAdvice(isConstructor(), packageName + ".JpmsClearanceAdvice");
   }
 }
