+++ dd-java-agent/instrumentation/mule-4/src/main/java11/datadog/trace/instrumentation/mule4/JpmsAdvisingHelper.java
@@ -1,22 +1,11 @@
 package datadog.trace.instrumentation.mule4;
 
-import java.util.WeakHashMap;
+import datadog.trace.api.GenericClassValue;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 public class JpmsAdvisingHelper {
-  private static final WeakHashMap<Class<?>, Boolean> ALREADY_PROCESSED_CACHE = new WeakHashMap<>();
-
-  public static void allowAccessOnModuleClass(final Class<?> cls) {
-    if (Boolean.TRUE.equals(ALREADY_PROCESSED_CACHE.putIfAbsent(cls, Boolean.TRUE))) {
-      return;
-    }
-    final Module module = cls.getModule();
-    if (module != null) {
-      try {
-        module.addExports(cls.getPackageName(), module.getClassLoader().getUnnamedModule());
-      } catch (Throwable ignored) {
-      }
-    }
-  }
+  public static final ClassValue<AtomicBoolean> ALREADY_PROCESSED_CACHE =
+      GenericClassValue.constructing(AtomicBoolean.class);
 
   private JpmsAdvisingHelper() {}
 }
